+++ source/sites/publicodes/Simulateur.js
@@ -16,7 +16,7 @@ import CarbonImpact from './CarbonImpact'
 import withTarget from './withTarget'
 import Chart from './chart/index.js'
 import { Redirect } from 'react-router'
-import PreviousSimulationBanner from 'Components/PreviousSimulationBanner'
+import SessionBar from 'Components/SessionBar'
 
 let CarbonImpactWithData = withTarget(CarbonImpact)
 
@@ -43,7 +43,7 @@ const Simulateur = (props) => {
 					<meta name="description" content={rule.description} />
 				)}
 			</Helmet>
-			<PreviousSimulationBanner />
+			<SessionBar />
 			<Simulation
 				noFeedback
 				noProgressMessage
